requirements.txt
paho-mqtt
PyYAML
grpcio
grpcio_tools